update opencode db symlink

+++ src/nix/home-manager/modules/activation/mutable-config/opencode.nix
@@ -11,6 +11,8 @@ helpers.mkSetupModule {
   script =
     { config, ... }:
     ''
-      ln -s "${config.home.homeDirectory}/.local/share/opencode/opencode-local.db" "${config.home.homeDirectory}/.local/share/opencode/opencode.db" 2>&1 >/dev/null
+      if [ -f "${config.home.homeDirectory}/.local/share/opencode/opencode.db" ]; then
+        ln -s "${config.home.homeDirectory}/.local/share/opencode/opencode-local.db" "${config.home.homeDirectory}/.local/share/opencode/opencode.db" 2>&1 >/dev/null
+      fi
     '';
 }
